dev-requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=dev-requirements.txt dev-requirements.in
#
alabaster==0.7.12
    # via sphinx
attrs==22.1.0
    # via pytest
babel==2.11.0
    # via sphinx
black==22.10.0
    # via -r dev-requirements.in
bleach==5.0.1
    # via readme-renderer
bump2version==1.0.1
    # via -r dev-requirements.in
certifi==2022.9.24
    # via
    #   -c requirements.txt
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via
    #   -c requirements.txt
    #   requests
click==8.1.3
    # via black
commonmark==0.9.1
    # via rich
coverage[toml]==6.5.0
    # via pytest-cov
cryptography==38.0.3
    # via secretstorage
distlib==0.3.6
    # via virtualenv
django-environ==0.9.0
    # via -r dev-requirements.in
docutils==0.19
    # via
    #   readme-renderer
    #   sphinx
exceptiongroup==1.0.4
    # via pytest
filelock==3.8.0
    # via
    #   tox
    #   virtualenv
freezegun==1.2.2
    # via -r dev-requirements.in
idna==3.4
    # via
    #   -c requirements.txt
    #   requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==5.0.0
    # via
    #   keyring
    #   twine
importlib-resources==5.10.0
    # via tox-gh-actions
iniconfig==1.1.1
    # via pytest
jaraco-classes==3.2.3
    # via keyring
jeepney==0.8.0
    # via
    #   keyring
    #   secretstorage
jinja2==3.1.2
    # via sphinx
keyring==23.11.0
    # via twine
markupsafe==2.1.1
    # via jinja2
more-itertools==9.0.0
    # via jaraco-classes
mypy-extensions==0.4.3
    # via black
packaging==21.3
    # via
    #   pytest
    #   sphinx
    #   tox
paho-mqtt==1.6.1
    # via -r dev-requirements.in
pathspec==0.10.2
    # via black
pkginfo==1.8.3
    # via twine
platformdirs==2.5.4
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via
    #   pytest
    #   tox
py==1.11.0
    # via tox
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via
    #   readme-renderer
    #   rich
    #   sphinx
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via
    #   -r dev-requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
pytest-cov==4.0.0
    # via -r dev-requirements.in
pytest-django==4.5.2
    # via -r dev-requirements.in
pytest-mock==3.10.0
    # via -r dev-requirements.in
python-dateutil==2.8.2
    # via freezegun
pytz==2022.6
    # via
    #   -c requirements.txt
    #   babel
readme-renderer==37.3
    # via twine
requests==2.28.1
    # via
    #   -c requirements.txt
    #   requests-toolbelt
    #   sphinx
    #   twine
requests-toolbelt==0.10.1
    # via twine
rfc3986==2.0.0
    # via twine
rich==12.6.0
    # via twine
secretstorage==3.3.3
    # via keyring
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
    #   tox
snowballstemmer==2.2.0
    # via sphinx
sphinx==5.3.0
    # via -r dev-requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pytest
    #   tox
tox==3.27.1
    # via
    #   -r dev-requirements.in
    #   tox-gh-actions
tox-gh-actions==2.11.0
    # via -r dev-requirements.in
twine==4.0.1
    # via -r dev-requirements.in
urllib3==1.26.12
    # via
    #   -c requirements.txt
    #   requests
    #   twine
virtualenv==20.16.7
    # via tox
webencodings==0.5.1
    # via bleach
wheel==0.38.4
    # via -r dev-requirements.in
zipp==3.10.0
    # via importlib-metadata
